app/requirements.txt
asgiref==3.4.1
Django==3.2.8
django-filter==21.1
django-tables2==2.4.1
djangorestframework==3.12.4
et-xmlfile==1.1.0
greenlet==1.1.2
Markdown==3.3.4
mysqlclient==2.0.3
numpy==1.21.4
openpyxl==3.0.9
pandas==1.3.4
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.26
sqlparse==0.4.2
